data-sources/bioRxiv/requirements_arxiv.txt
feedparser==6.0.11
Flask==3.0.3
numpy==2.1.2
pandas==2.2.3
python-telegram-bot==21.6
Requests==2.32.3
tqdm==4.66.1
psycopg2-binary==2.9.9
ollama==0.3.3
python-dotenv==1.0.1
